Move processing into pipeline refactoring.
